// web/src/lib/types/domain.ts
export type Pais = 'NI' | 'CR' | 'HN';
export type SemaforoNivel = 'V' | 'A' | 'R';
// ...
export interface Paciente {
  id_paciente?: number;
  carnet: string;
  nombre_completo: string;
  fecha_nacimiento?: string;
  sexo?: string;
  telefono?: string;
  correo?: string;
  gerencia?: string;
  area?: string;
  estado_paciente: 'A' | 'I';
  nivel_semaforo?: SemaforoNivel;
  pais?: Pais;
  correo_usuario?: string;
  // Aliases para compatibilidad
  id?: number;
  nombreCompleto?: string;
}